automatically backed up by learn

+++ index.js
@@ -23,5 +23,5 @@ function dec(a) {
 }
 
 function  makeInt(n) {
-return (parseInt ('n',100))
+return (parseFloat ('n',100))
 }
